Move stored-page classification into lib beside its write-time half
